Identify unused switch ports by parsing logs
Can I unplug this switch port?

It’s a pretty common problem.

If you have logs with link state, and can get the current state of the
interfaces, then yes, you can identify the last time a switch port was
used, and then decide if you want to unplug it  and use the port for
another jack.